test(itest): split solstice_upgrade_quality_test.go to avoid CI timeout
The original file contained 5 test functions and consistently exceeded
the 10-minute CI time limit. Split into three separate files so each
gets its own CI job slot:

- solstice_upgrade_quality_auth_test.go: UpgradeQualityAuth + PureOwnerAuth
- solstice_fault_recover_test.go: FaultAndRecover + FaultRecoverFullPower
  (+ shared helper waitFaultedAndPastDeadline)
- solstice_fault_recover_usqd_test.go: FaultRecoverUsqdFullPower
